+++ src/bot/events.rs
@@ -1,6 +1,4 @@
 //! This module contains all the bevy events that are send
-//!
-//! NOTE: Every Event has a prefix `B` to avoid confusion with `serenity` events.
 
 use bevy_ecs::prelude::*;
 use serenity::all::*;
@@ -10,7 +8,7 @@ use std::collections::HashMap;
 /// Dispatched upon startup.
 ///
 /// Provides data about the bot and the guilds it’s in.
-pub struct BReadyEvent {
+pub struct ReadyEvent {
     pub ctx: Context,
     pub data_about_bot: Ready,
 }
@@ -19,7 +17,7 @@ pub struct BReadyEvent {
 /// Dispatched when the permissions of an application command was updated.
 ///
 /// Provides said permission’s data.
-pub struct BCommandPermissionsUpdate {
+pub struct CommandPermissionsUpdate {
     pub ctx: Context,
     pub permission: CommandPermissions,
 }
@@ -28,7 +26,7 @@ pub struct BCommandPermissionsUpdate {
 /// Dispatched when an auto moderation rule was created.
 ///
 /// Provides said rule’s data.
-pub struct BAutoModerationRuleCreate {
+pub struct AutoModerationRuleCreate {
     pub ctx: Context,
     pub rule: Rule,
 }
@@ -37,7 +35,7 @@ pub struct BAutoModerationRuleCreate {
 /// Dispatched when an auto moderation rule was updated.
 ///
 /// Provides said rule’s data.
-pub struct BAutoModerationRuleUpdate {
+pub struct AutoModerationRuleUpdate {
     pub ctx: Context,
     pub rule: Rule,
 }
@@ -46,7 +44,7 @@ pub struct BAutoModerationRuleUpdate {
 /// Dispatched when an auto moderation rule was deleted.
 ///
 /// Provides said rule’s data.
-pub struct BAutoModerationRuleDelete {
+pub struct AutoModerationRuleDelete {
     pub ctx: Context,
     pub rule: Rule,
 }
@@ -55,7 +53,7 @@ pub struct BAutoModerationRuleDelete {
 /// Dispatched when an auto moderation rule was triggered and an action was executed.
 ///
 /// Provides said action execution’s data.
-pub struct BAutoModerationActionExecution {
+pub struct AutoModerationActionExecution {
     pub ctx: Context,
     pub execution: ActionExecution,
 }
@@ -69,7 +67,7 @@ pub struct BAutoModerationActionExecution {
 /// performed prior this event may fail as the data could be not inserted yet.
 ///
 /// Provides the cached guilds’ ids.
-pub struct BCacheRead {
+pub struct CacheRead {
     pub ctx: Context,
     pub guilds: Vec<GuildId>,
 }
@@ -78,7 +76,7 @@ pub struct BCacheRead {
 #[cfg_attr(docsrs, doc(cfg(feature = "bot_cache")))]
 #[derive(Event)]
 /// Dispatched when every shard has received a Ready event
-pub struct BShardsReady {
+pub struct ShardsReady {
     pub ctx: Context,
     pub total_shards: u32,
 }
@@ -87,7 +85,7 @@ pub struct BShardsReady {
 /// Dispatched when a channel is created.
 ///
 /// Provides said channel’s data.
-pub struct BChannelCreate {
+pub struct ChannelCreate {
     pub ctx: Context,
     pub channel: GuildChannel,
 }
@@ -96,7 +94,7 @@ pub struct BChannelCreate {
 /// Dispatched when a category is created.
 ///
 /// Provides said category’s data.
-pub struct BCategoryCreate {
+pub struct CategoryCreate {
     pub ctx: Context,
     pub category: GuildChannel,
 }
@@ -105,7 +103,7 @@ pub struct BCategoryCreate {
 /// Dispatched when a category is deleted.
 ///
 /// Provides said category’s data.
-pub struct BCategoryDelete {
+pub struct CategoryDelete {
     pub ctx: Context,
     pub category: GuildChannel,
 }
@@ -114,7 +112,7 @@ pub struct BCategoryDelete {
 /// Dispatched when a channel is deleted.
 ///
 /// Provides said channel’s data.
-pub struct BChannelDelete {
+pub struct ChannelDelete {
     pub ctx: Context,
     pub channel: GuildChannel,
     pub messages: Option<Vec<Message>>,
@@ -124,7 +122,7 @@ pub struct BChannelDelete {
 /// Dispatched when a pin is added, deleted.
 ///
 /// Provides said pin’s data.
-pub struct BChannelPinUpdate {
+pub struct ChannelPinUpdate {
     pub ctx: Context,
     pub pin: ChannelPinsUpdateEvent,
 }
@@ -133,7 +131,7 @@ pub struct BChannelPinUpdate {
 /// Dispatched when a channel is updated.
 ///
 /// The old channel data is only provided when the `bot_cache` feature is enabled.
-pub struct BChannelUpdate {
+pub struct ChannelUpdate {
     pub ctx: Context,
     pub old: Option<GuildChannel>,
     pub new: GuildChannel,
@@ -143,7 +141,7 @@ pub struct BChannelUpdate {
 /// Dispatched when a new audit log entry is created.
 ///
 /// Provides said entry’s data and the id of the guild where it was created.
-pub struct BGuildAuditLogEntryCreate {
+pub struct GuildAuditLogEntryCreate {
     pub ctx: Context,
     pub entry: AuditLogEntry,
     pub guild_id: GuildId,
@@ -153,7 +151,7 @@ pub struct BGuildAuditLogEntryCreate {
 /// Dispatched when a user is banned from a guild.
 ///
 /// Provides the guild’s id and the banned user’s data.
-pub struct BGuildBanAddition {
+pub struct GuildBanAddition {
     pub ctx: Context,
     pub guild_id: GuildId,
     pub banned_user: User,
@@ -163,7 +161,7 @@ pub struct BGuildBanAddition {
 /// Dispatched when a user’s ban is lifted from a guild.
 ///
 /// Provides the guild’s id and the lifted user’s data.
-pub struct BGuildBanRemoval {
+pub struct GuildBanRemoval {
     pub ctx: Context,
     pub guild_id: GuildId,
     pub unbanned_user: User,
@@ -173,7 +171,7 @@ pub struct BGuildBanRemoval {
 /// Dispatched when a guild is created; or an existing guild’s data is sent to us.
 ///
 /// Provides the guild’s data and whether the guild is new (only when `bot_cache` feature is enabled).
-pub struct BGuildCreate {
+pub struct GuildCreate {
     pub ctx: Context,
     pub guild: Guild,
     pub is_new: Option<bool>,
@@ -188,7 +186,7 @@ pub struct BGuildCreate {
 /// The [`UnavailableGuild::unavailable`] flag in the partial data determines the status of the guild. If the flag
 /// is false, the bot was removed from the guild, either by being kicked or banned. If the
 /// flag is true, the guild went offline.
-pub struct BGuildDelete {
+pub struct GuildDelete {
     pub ctx: Context,
     pub incomplete: UnavailableGuild,
     pub full: Option<Guild>,
@@ -198,7 +196,7 @@ pub struct BGuildDelete {
 /// Dispatched when the emojis are updated.
 ///
 /// Provides the guild’s id and the new state of the emojis in the guild.
-pub struct BGuildEmojisUpdate {
+pub struct GuildEmojisUpdate {
     pub ctx: Context,
     pub guild_id: GuildId,
     pub current_state: HashMap<EmojiId, Emoji>,
@@ -208,7 +206,7 @@ pub struct BGuildEmojisUpdate {
 /// Dispatched when a guild’s integration is added, updated or removed.
 ///
 /// Provides the guild’s id.
-pub struct BGuildIntegrationsUpdate {
+pub struct GuildIntegrationsUpdate {
     pub ctx: Context,
     pub guild_id: GuildId,
 }
@@ -220,7 +218,7 @@ pub struct BGuildIntegrationsUpdate {
 ///
 /// Note: This event will not trigger unless the “guild members” privileged intent is enabled
 /// on the bot application page.
-pub struct BGuildMemberAddition {
+pub struct GuildMemberAddition {
     pub ctx: Context,
     pub new_member: Member,
 }
@@ -233,7 +231,7 @@ pub struct BGuildMemberAddition {
 ///
 /// Note: This event will not trigger unless the “guild members” privileged intent is enabled
 /// on the bot application page.
-pub struct BGuildMemberRemoval {
+pub struct GuildMemberRemoval {
     pub ctx: Context,
     pub guild_id: GuildId,
     pub user: User,
@@ -248,7 +246,7 @@ pub struct BGuildMemberRemoval {
 ///
 /// Note: This event will not trigger unless the “guild members” privileged intent is enabled
 /// on the bot application page.
-pub struct BGuildMemberUpdate {
+pub struct GuildMemberUpdate {
     pub ctx: Context,
     pub old_if_available: Option<Member>,
     pub new: Option<Member>,
@@ -259,7 +257,7 @@ pub struct BGuildMemberUpdate {
 /// Dispatched when the data for offline members was requested.
 ///
 /// Provides the guild’s id and the data.
-pub struct BGuildMembersChunk {
+pub struct GuildMembersChunk {
     pub ctx: Context,
     pub chunk: GuildMembersChunkEvent,
 }
@@ -268,7 +266,7 @@ pub struct BGuildMembersChunk {
 /// Dispatched when a role is created.
 ///
 /// Provides the guild’s id and the new role’s data.
-pub struct BGuildRoleCreate {
+pub struct GuildRoleCreate {
     pub ctx: Context,
     pub new: Role,
 }
@@ -277,7 +275,7 @@ pub struct BGuildRoleCreate {
 /// Dispatched when a role is deleted.
 /// Provides the guild’s id, the role’s id and its data (if `bot_cache` feature is enabled
 /// and the data is available).
-pub struct BGuildRoleDelete {
+pub struct GuildRoleDelete {
     pub ctx: Context,
     pub guild_id: GuildId,
     pub removed_role_id: RoleId,
@@ -289,7 +287,7 @@ pub struct BGuildRoleDelete {
 ///
 /// Provides the guild’s id, the role’s old (if `bot_cache` feature is enabled and the data
 /// is available) and new data.
-pub struct BGuildRoleUpdate {
+pub struct GuildRoleUpdate {
     pub ctx: Context,
     pub old_data_if_available: Option<Role>,
     pub new: Role,
@@ -299,7 +297,7 @@ pub struct BGuildRoleUpdate {
 /// Dispatched when the stickers are updated.
 ///
 /// Provides the guild’s id and the new state of the stickers in the guild.
-pub struct BGuildStickersUpdate {
+pub struct GuildStickersUpdate {
     pub ctx: Context,
     pub guild_id: GuildId,
     pub current_state: HashMap<StickerId, Sticker>,
@@ -310,7 +308,7 @@ pub struct BGuildStickersUpdate {
 ///
 /// Provides the guild’s old data (if `bot_cache` feature is enabled and the data is
 /// available) and the new data.
-pub struct BGuildUpdate {
+pub struct GuildUpdate {
     pub ctx: Context,
     pub old_data_if_available: Option<Guild>,
     pub new_data: PartialGuild,
@@ -320,7 +318,7 @@ pub struct BGuildUpdate {
 /// Dispatched when a invite is created.
 ///
 /// Provides data about the invite.
-pub struct BInviteCreate {
+pub struct InviteCreate {
     pub ctx: Context,
     pub data: InviteCreateEvent,
 }
@@ -329,7 +327,7 @@ pub struct BInviteCreate {
 /// Dispatched when a invite is deleted.
 ///
 /// Provides data about the invite.
-pub struct BInviteDelete {
+pub struct InviteDelete {
     pub ctx: Context,
     pub data: InviteDeleteEvent,
 }
@@ -338,7 +336,7 @@ pub struct BInviteDelete {
 /// Dispatched when a message is created.
 ///
 /// Provides the message’s data.
-pub struct BMessage {
+pub struct Message {
     pub ctx: Context,
     pub new_message: Message,
 }
@@ -347,7 +345,7 @@ pub struct BMessage {
 /// Dispatched when a message is deleted.
 ///
 /// Provides the guild’s id, the channel’s id and the message’s id.
-pub struct BMessageDelete {
+pub struct MessageDelete {
     pub ctx: Context,
     pub channel_id: ChannelId,
     pub deleted_message_id: MessageId,
@@ -358,7 +356,7 @@ pub struct BMessageDelete {
 /// Dispatched when multiple messages were deleted at once.
 ///
 /// Provides the guild’s id, channel’s id and the deleted messages’ ids.
-pub struct BMessageDeleteBulk {
+pub struct MessageDeleteBulk {
     pub ctx: Context,
     pub channel_id: ChannelId,
     pub multiple_deleted_messages_ids: Vec<MessageId>,
@@ -370,7 +368,7 @@ pub struct BMessageDeleteBulk {
 ///
 /// Provides the message update data, as well as the actual old and new message if `bot_cache`
 /// feature is enabled and the data is available.
-pub struct BMessageUpdate {
+pub struct MessageUpdate {
     pub ctx: Context,
     pub old_if_available: Option<Message>,
     pub new: Option<Message>,
@@ -381,7 +379,7 @@ pub struct BMessageUpdate {
 /// Dispatched when a new reaction is attached to a message.
 ///
 /// Provides the reaction’s data.
-pub struct BReactionAdd {
+pub struct ReactionAdd {
     pub ctx: Context,
     pub add_reaction: Reaction,
 }
@@ -390,7 +388,7 @@ pub struct BReactionAdd {
 /// Dispatched when a reaction is detached from a message.
 ///
 /// Provides the reaction’s data.
-pub struct BReactionRemove {
+pub struct ReactionRemove {
     pub ctx: Context,
     pub removed_reaction: Reaction,
 }
@@ -399,7 +397,7 @@ pub struct BReactionRemove {
 /// Dispatched when all reactions of a message are detached from a message.
 ///
 /// Provides the channel’s id and the message’s id.
-pub struct BReactionRemoveAll {
+pub struct ReactionRemoveAll {
     pub ctx: Context,
     pub channel_id: ChannelId,
     pub removed_from_message_id: MessageId,
@@ -409,7 +407,7 @@ pub struct BReactionRemoveAll {
 /// Dispatched when all reactions of a message are detached from a message.
 ///
 /// Provides the channel’s id and the message’s id.
-pub struct BReactionRemoveEmoji {
+pub struct ReactionRemoveEmoji {
     pub ctx: Context,
     pub removed_reactions: Reaction,
 }
@@ -421,14 +419,14 @@ pub struct BReactionRemoveEmoji {
 ///
 /// Note: This event will not trigger unless the “guild presences” privileged intent is enabled
 /// on the bot application page.
-pub struct BPresenceUpdate {
+pub struct PresenceUpdate {
     pub ctx: Context,
     pub new_data: Presence,
 }
 
 #[derive(Event)]
 /// Dispatched upon reconnection.
-pub struct BResume {
+pub struct Resume {
     pub ctx: Context,
     pub event: ResumedEvent,
 }
@@ -437,14 +435,14 @@ pub struct BResume {
 /// Dispatched when a shard’s connection stage is updated
 ///
 /// Provides the context of the shard and the event information about the update.
-pub struct BShardStageUpdate {
+pub struct ShardStageUpdate {
     pub ctx: Context,
     pub event: ShardStageUpdateEvent,
 }
 
 #[derive(Event)]
 /// Dispatched when a user starts typing.
-pub struct BTypingStart {
+pub struct TypingStart {
     pub ctx: Context,
     pub event: TypingStartEvent,
 }
@@ -453,7 +451,7 @@ pub struct BTypingStart {
 /// Dispatched when the bot’s data is updated.
 ///
 /// Provides the old (if `bot_cache` feature is enabled and the data is available) and new data.
-pub struct BUserUpdate {
+pub struct UserUpdate {
     pub ctx: Context,
     pub old_data: Option<CurrentUser>,
     pub new: CurrentUser,
@@ -463,7 +461,7 @@ pub struct BUserUpdate {
 /// Dispatched when a guild’s voice server was updated (or changed to another one).
 ///
 /// Provides the voice server’s data.
-pub struct BVoiceServerUpdate {
+pub struct VoiceServerUpdate {
     pub ctx: Context,
     pub event: VoiceServerUpdateEvent,
 }
@@ -473,7 +471,7 @@ pub struct BVoiceServerUpdate {
 ///
 /// Provides the guild’s id (if available) and the old state (if `bot_cache` feature is enabled
 /// and [`GatewayIntents::GUILDS`] is enabled) and the new state of the guild’s voice channels.
-pub struct BVoiceStateUpdate {
+pub struct VoiceStateUpdate {
     pub ctx: Context,
     pub old: Option<VoiceState>,
     pub new: VoiceState,
@@ -483,7 +481,7 @@ pub struct BVoiceStateUpdate {
 /// Dispatched when a voice channel’s status is updated.
 ///
 /// Provides the status, channel’s id and the guild’s id.
-pub struct BVoiceChannelStatusUpdate {
+pub struct VoiceChannelStatusUpdate {
     pub ctx: Context,
     pub old: Option<String>,
     pub status: Option<String>,
@@ -495,7 +493,7 @@ pub struct BVoiceChannelStatusUpdate {
 /// Dispatched when a guild’s webhook is updated.
 ///
 /// Provides the guild’s id and the channel’s id the webhook belongs in.
-pub struct BWebhookUpdate {
+pub struct WebhookUpdate {
     pub ctx: Context,
     pub guild_id: GuildId,
     pub belongs_to_channel_id: ChannelId,
@@ -506,7 +504,7 @@ pub struct BWebhookUpdate {
 /// clicked).
 ///
 /// Provides the created interaction.
-pub struct BInteractionCreate {
+pub struct InteractionCreate {
     pub ctx: Context,
     pub interaction: Interaction,
 }
@@ -515,7 +513,7 @@ pub struct BInteractionCreate {
 /// Dispatched when a guild integration is created.
 ///
 /// Provides the created integration.
-pub struct BIntegrationCreate {
+pub struct IntegrationCreate {
     pub ctx: Context,
     pub integration: Integration,
 }
@@ -524,7 +522,7 @@ pub struct BIntegrationCreate {
 /// Dispatched when a guild integration is updated.
 ///
 /// Provides the updated integration.
-pub struct BIntegrationUpdate {
+pub struct IntegrationUpdate {
     pub ctx: Context,
     pub integration: Integration,
 }
@@ -533,7 +531,7 @@ pub struct BIntegrationUpdate {
 /// Dispatched when a stage instance is created.
 ///
 /// Provides the created stage instance.
-pub struct BStageInstanceCreate {
+pub struct StageInstanceCreate {
     pub ctx: Context,
     pub stage_instance: StageInstance,
 }
@@ -542,7 +540,7 @@ pub struct BStageInstanceCreate {
 /// Dispatched when a stage instance is updated.
 ///
 /// Provides the updated stage instance.
-pub struct BStageInstanceUpdate {
+pub struct StageInstanceUpdate {
     pub ctx: Context,
     pub stage_instance: StageInstance,
 }
@@ -551,7 +549,7 @@ pub struct BStageInstanceUpdate {
 /// Dispatched when a stage instance is deleted.
 ///
 /// Provides the deleted stage instance.
-pub struct BStageInstanceDelete {
+pub struct StageInstanceDelete {
     pub ctx: Context,
     pub stage_instance: StageInstance,
 }
@@ -560,7 +558,7 @@ pub struct BStageInstanceDelete {
 /// Dispatched when a thread is created or the current user is added to a private thread.
 ///
 /// Provides the thread.
-pub struct BThreadCreate {
+pub struct ThreadCreate {
     pub ctx: Context,
     pub thread: GuildChannel,
 }
@@ -570,7 +568,7 @@ pub struct BThreadCreate {
 ///
 /// Provides the updated thread and the old thread data, provided the thread was cached prior
 /// to dispatch.
-pub struct BThreadUpdate {
+pub struct ThreadUpdate {
     pub ctx: Context,
     pub old: Option<GuildChannel>,
     pub new: GuildChannel,
@@ -581,7 +579,7 @@ pub struct BThreadUpdate {
 ///
 /// Provides the partial data about the deleted thread and, if it was present in the cache
 /// before its deletion, its full data.
-pub struct BThreadDelete {
+pub struct ThreadDelete {
     pub ctx: Context,
     pub thread: PartialGuildChannel,
     pub full_thread_data: Option<GuildChannel>,
@@ -592,7 +590,7 @@ pub struct BThreadDelete {
 ///
 /// Provides the threads the current user can access, the thread members, the guild Id,
 /// and the channel Ids of the parent channels being synced.
-pub struct BThreadListSync {
+pub struct ThreadListSync {
     pub ctx: Context,
     pub thread_list_sync: ThreadListSyncEvent,
 }
@@ -601,7 +599,7 @@ pub struct BThreadListSync {
 /// Dispatched when the [`ThreadMember`] for the current user is updated.
 ///
 /// Provides the updated thread member.
-pub struct BThreadMemberUpdate {
+pub struct ThreadMemberUpdate {
     pub ctx: Context,
     pub thread_member: ThreadMember,
 }
@@ -613,7 +611,7 @@ pub struct BThreadMemberUpdate {
 ///
 /// Provides the added/removed members, the approximate member count of members in the thread,
 /// the thread Id and its guild Id.
-pub struct BThreadMembersUpdate {
+pub struct ThreadMembersUpdate {
     pub ctx: Context,
     pub thread_members_update: ThreadMembersUpdateEvent,
 }
@@ -622,7 +620,7 @@ pub struct BThreadMembersUpdate {
 /// Dispatched when a scheduled event is created.
 ///
 /// Provides data about the scheduled event.
-pub struct BGuildScheduledEventCreate {
+pub struct GuildScheduledEventCreate {
     pub ctx: Context,
     pub event: ScheduledEvent,
 }
@@ -631,7 +629,7 @@ pub struct BGuildScheduledEventCreate {
 /// Dispatched when a scheduled event is updated.
 ///
 /// Provides data about the scheduled event.
-pub struct BGuildScheduledEventUpdate {
+pub struct GuildScheduledEventUpdate {
     pub ctx: Context,
     pub event: ScheduledEvent,
 }
@@ -640,7 +638,7 @@ pub struct BGuildScheduledEventUpdate {
 /// Dispatched when a scheduled event is deleted.
 ///
 /// Provides data about the scheduled event.
-pub struct BGuildScheduledEventDelete {
+pub struct GuildScheduledEventDelete {
     pub ctx: Context,
     pub event: ScheduledEvent,
 }
@@ -649,7 +647,7 @@ pub struct BGuildScheduledEventDelete {
 /// Dispatched when a guild member has subscribed to a scheduled event.
 ///
 /// Provides data about the subscription.
-pub struct BGuildScheduledEventUserAdd {
+pub struct GuildScheduledEventUserAdd {
     pub ctx: Context,
     pub subscribed: GuildScheduledEventUserAddEvent,
 }
@@ -658,7 +656,7 @@ pub struct BGuildScheduledEventUserAdd {
 /// Dispatched when a guild member has unsubscribed from a scheduled event.
 ///
 /// Provides data about the cancelled subscription.
-pub struct BGuildScheduledEventUserRemove {
+pub struct GuildScheduledEventUserRemove {
     pub ctx: Context,
     pub unsubscribed: GuildScheduledEventUserRemoveEvent,
 }
@@ -667,7 +665,7 @@ pub struct BGuildScheduledEventUserRemove {
 /// Dispatched when a user subscribes to a SKU.
 ///
 /// Provides data about the subscription.
-pub struct BEntitlementCreate {
+pub struct EntitlementCreate {
     pub ctx: Context,
     pub entitlement: Entitlement,
 }
@@ -678,7 +676,7 @@ pub struct BEntitlementCreate {
 ///
 /// Provides data abut the updated subscription. If the entitlement is renewed, the
 /// `[Entitlement::ends_at`] field will have changed.
-pub struct BEntitlementUpdate {
+pub struct EntitlementUpdate {
     pub ctx: Context,
     pub entitlement: Entitlement,
 }
@@ -689,7 +687,7 @@ pub struct BEntitlementUpdate {
 /// when they expire.
 ///
 /// Provides data about the subscription. Specifically, the Entitlement::deleted field will be set.
-pub struct BEntitlementDelete {
+pub struct EntitlementDelete {
     pub ctx: Context,
     pub entitlement: Entitlement,
 }
@@ -698,20 +696,20 @@ pub struct BEntitlementDelete {
 /// Dispatched when a user votes on a message poll.
 ///
 /// This will be dispatched multiple times if multiple answers are selected.
-pub struct BPollVoteAdd {
+pub struct PollVoteAdd {
     pub ctx: Context,
     pub event: MessagePollVoteAddEvent,
 }
 
 #[derive(Event)]
 /// Dispatched when a user removes a previous vote on a poll.
-pub struct BPollVoteRemove {
+pub struct PollVoteRemove {
     pub ctx: Context,
     pub event: MessagePollVoteRemoveEvent,
 }
 
 #[derive(Event)]
 /// Dispatched when an HTTP rate limit is hit
-pub struct BRateLimit {
+pub struct RateLimit {
     pub data: RatelimitInfo,
 }